\name{gevcdn.identity}
\alias{gevcdn.identity}
\title{
Identity function
}
\description{
\code{gevcdn.identity} computes a trivial identity function. Used as the hidden layer transfer function for linear GEV CDN models.
}
\usage{
gevcdn.identity(x)
}

\arguments{
       \item{x}{
           a numeric vector
       }
}

\seealso{
    \code{\link{gevcdn.logistic}}
}
