\name{gevcdn.logistic}
\alias{gevcdn.logistic}
\title{
Logistic sigmoid function
}
\description{
\code{gevcdn.logistic} computes the logistic sigmoid (S-shaped) function. Used as the hidden layer transfer function for nonlinear GEV CDN models.
}
\usage{
gevcdn.logistic(x)
}

\arguments{
       \item{x}{
           a numeric vector
       }
}

\seealso{
    \code{\link{gevcdn.identity}}
}
